requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7
python-dotenv==1.0.0
jinja2
supabase==2.8.1
httpx==0.28.1
python-multipart==0.0.6
email-validator==2.0.0
